+++ src/http/control/mod.rs
@@ -284,6 +284,8 @@ mod tests {
             tcp_ws_mode: WsTransportMode::Http1,
             udp_ws_url: None,
             udp_ws_mode: WsTransportMode::Http1,
+            vless_ws_url: None,
+            vless_ws_mode: WsTransportMode::Http1,
             tcp_addr: Some(addr.to_string().parse::<ServerAddr>().unwrap()),
             udp_addr: None,
             cipher: CipherKind::Chacha20IetfPoly1305,